WI #17 Move components to components folder
